+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added possibility to change certain strings/logos in permission and login flow layers
 - Added possibility to get back buttons to build an own authorization dialog
 - Added LSApplicationQueriesSchemes in Info.plist
+- Added demo app for button worflow
 
 ### Changed
 - Added more information to user info